+++ encoder_test.go
@@ -7,17 +7,585 @@ package waf
 
 import (
 	"testing"
+	"time"
+	"unsafe"
 
-	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-func TestEncodeNilMap(t *testing.T) {
-	var nilMap map[string]any = nil
+func TestEncoder(t *testing.T) {
+	for _, tc := range []struct {
+		Name                   string
+		Data                   interface{}
+		ExpectedError          error
+		ExpectedWAFValueType   int
+		ExpectedWAFValueLength int
+		ExpectedWAFString      string
+		ExpectedWafValueInt    int64
+		ExpectedWafValueUint   uint64
+		ExpectedWafValueFloat  float64
+		ExpectedWafValueBool   bool
+		MaxValueDepth          interface{}
+		MaxContainerLength     interface{}
+		MaxStringLength        interface{}
+	}{
+		{
+			Name:          "unsupported type",
+			Data:          make(chan struct{}),
+			ExpectedError: errUnsupportedValue,
+		},
+		{
+			Name:                 "string-hekki",
+			Data:                 "hello, waf",
+			ExpectedWAFValueType: wafStringType,
+			ExpectedWAFString:    "hello, waf",
+		},
+		{
+			Name:                   "string-empty",
+			Data:                   "",
+			ExpectedWAFValueType:   wafStringType,
+			ExpectedWAFValueLength: 0,
+		},
+		{
+			Name:                 "byte-slice",
+			Data:                 []byte("hello, waf"),
+			ExpectedWAFValueType: wafStringType,
+			ExpectedWAFString:    "hello, waf",
+		},
+		{
+			Name:                   "nil-byte-slice",
+			Data:                   []byte(nil),
+			ExpectedWAFValueType:   wafStringType,
+			ExpectedWAFValueLength: 0,
+		},
+		{
+			Name:                   "nil-map",
+			Data:                   map[string]any(nil),
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 0,
+		},
+		{
+			Name:                   "map-with-empty-key-string",
+			Data:                   map[string]int{"": 1},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 1,
+		},
+		{
+			Name:                   "empty-struct",
+			Data:                   struct{}{},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 0,
+		},
+		{
+			Name: "empty-struct-with-private-fields",
+			Data: struct {
+				a string
+				b int
+				c bool
+			}{},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 0,
+		},
+		{
+			Name:          "nil-interface-value",
+			Data:          nil,
+			ExpectedError: errUnsupportedValue,
+		},
+		{
+			Name:          "nil-pointer-value",
+			Data:          (*string)(nil),
+			ExpectedError: errUnsupportedValue,
+		},
+		{
+			Name:          "nil-pointer-value",
+			Data:          (*int)(nil),
+			ExpectedError: errUnsupportedValue,
+		},
+		{
+			Name:                 "non-nil-pointer-value",
+			Data:                 new(int),
+			ExpectedWAFValueType: wafIntType,
+			ExpectedWafValueInt:  0,
+		},
+		{
+			Name:                   "non-nil-pointer-value",
+			Data:                   new(string),
+			ExpectedWAFValueType:   wafStringType,
+			ExpectedWAFValueLength: 0,
+		},
+		{
+			Name:                   "having-an-empty-map",
+			Data:                   map[string]interface{}{},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 0,
+		},
+		{
+			Name:          "unsupported",
+			Data:          func() {},
+			ExpectedError: errUnsupportedValue,
+		},
+		{
+			Name:                 "int",
+			Data:                 int(1234),
+			ExpectedWAFValueType: wafIntType,
+			ExpectedWafValueInt:  1234,
+		},
+		{
+			Name:                 "uint",
+			Data:                 uint(9876),
+			ExpectedWAFValueType: wafUintType,
+			ExpectedWafValueUint: 9876,
+		},
+		{
+			Name:                 "bool",
+			Data:                 true,
+			ExpectedWAFValueType: wafBoolType,
+			ExpectedWafValueBool: true,
+		},
+		{
+			Name:                 "bool",
+			Data:                 false,
+			ExpectedWAFValueType: wafBoolType,
+			ExpectedWafValueBool: false,
+		},
+		{
+			Name:                  "float",
+			Data:                  33.12345,
+			ExpectedWAFValueType:  wafFloatType,
+			ExpectedWafValueFloat: 33.12345,
+		},
+		{
+			Name:                  "float",
+			Data:                  33.62345,
+			ExpectedWAFValueType:  wafFloatType,
+			ExpectedWafValueFloat: 33.62345,
+		},
+		{
+			Name:                   "slice",
+			Data:                   []interface{}{33.12345, "ok", 27},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 3,
+		},
+		{
+			Name:                   "slice-having-unsupported-values",
+			Data:                   []interface{}{33.12345, func() {}, "ok", 27, nil},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 3,
+		},
+		{
+			Name:                   "array",
+			Data:                   [...]interface{}{func() {}, 33.12345, "ok", 27},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 3,
+		},
+		{
+			Name:                   "map",
+			Data:                   map[string]interface{}{"k1": 1, "k2": "2"},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 2,
+		},
+		{
+			Name:                   "map-with-unsupported-key-values",
+			Data:                   map[interface{}]interface{}{"k1": 1, 27: "int key", "k2": "2"},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 2,
+		},
+		{
+			Name:                   "map-with-indirect-key-string-values",
+			Data:                   map[interface{}]interface{}{"k1": 1, new(string): "string pointer key", "k2": "2"},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 3,
+		},
+		{
+			Name: "struct",
+			Data: struct {
+				Public  string
+				private string
+				a       string
+				A       string
+			}{
+				Public:  "Public",
+				private: "private",
+				a:       "a",
+				A:       "A",
+			},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 2, // public fields only
+		},
+		{
+			Name: "struct-with-unsupported-values",
+			Data: struct {
+				Public  string
+				private string
+				a       string
+				A       func()
+			}{
+				Public:  "Public",
+				private: "private",
+				a:       "a",
+				A:       nil,
+			},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 1, // public fields of supported types
+		},
+		{
+			Name:                   "array-max-depth",
+			MaxValueDepth:          0,
+			Data:                   []interface{}{1, 2, 3, 4, []int{1, 2, 3, 4}},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 4,
+		},
+		{
+			Name:                   "array-max-depth",
+			MaxValueDepth:          1,
+			Data:                   []interface{}{1, 2, 3, 4, []int{1, 2, 3, 4}},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 5,
+		},
+		{
+			Name:                   "array-max-depth",
+			MaxValueDepth:          0,
+			Data:                   []interface{}{},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 0,
+		},
+		{
+			Name:                   "map-max-depth",
+			MaxValueDepth:          0,
+			Data:                   map[string]interface{}{"k1": "v1", "k2": "v2", "k3": "v3", "k4": "v4", "k5": map[string]string{}},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 4,
+		},
+		{
+			Name:                   "map-max-depth",
+			MaxValueDepth:          1,
+			Data:                   map[string]interface{}{"k1": "v1", "k2": "v2", "k3": "v3", "k4": "v4", "k5": map[string]string{}},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 5,
+		},
+		{
+			Name:                   "map-max-depth",
+			MaxValueDepth:          0,
+			Data:                   map[string]interface{}{},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 0,
+		},
+		{
+			Name:                   "struct-max-depth",
+			MaxValueDepth:          0,
+			Data:                   struct{}{},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 0,
+		},
+		{
+			Name:          "struct-max-depth",
+			MaxValueDepth: 0,
+			Data: struct {
+				F0 string
+				F1 struct{}
+			}{},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 1,
+		},
+		{
+			Name:          "struct-max-depth",
+			MaxValueDepth: 1,
+			Data: struct {
+				F0 string
+				F1 struct{}
+			}{},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 2,
+		},
+		{
+			Name:                 "scalar-values-max-depth-not-accounted",
+			MaxValueDepth:        0,
+			Data:                 -1234,
+			ExpectedWafValueInt:  -1234,
+			ExpectedWAFValueType: wafIntType,
+		},
+		{
+			Name:                 "scalar-values-max-depth-not-accounted",
+			MaxValueDepth:        0,
+			Data:                 uint(1234),
+			ExpectedWafValueUint: 1234,
+			ExpectedWAFValueType: wafUintType,
+		},
+		{
+			Name:                 "scalar-values-max-depth-not-accounted",
+			MaxValueDepth:        0,
+			Data:                 false,
+			ExpectedWAFValueType: wafBoolType,
+			ExpectedWafValueBool: false,
+		},
+		{
+			Name:                   "array-max-length",
+			MaxContainerLength:     3,
+			Data:                   []interface{}{1, 2, 3, 4, 5},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 3,
+		},
+		{
+			Name:                   "map-max-length",
+			MaxContainerLength:     3,
+			Data:                   map[string]string{"k1": "v1", "k2": "v2", "k3": "v3", "k4": "v4", "k5": "v5"},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 3,
+		},
+		{
+			Name:                 "string-max-length",
+			MaxStringLength:      3,
+			Data:                 "123456789",
+			ExpectedWAFString:    "123",
+			ExpectedWAFValueType: wafStringType,
+		},
+		{
+			Name:                   "string-max-length-truncation-leading-to-same-map-keys",
+			MaxStringLength:        1,
+			Data:                   map[string]string{"k1": "v1", "k2": "v2", "k3": "v3", "k4": "v4", "k5": "v5"},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 5,
+		},
+		{
+			Name:                   "unsupported-array-values",
+			MaxContainerLength:     1,
+			Data:                   []interface{}{"supported", func() {}, "supported", make(chan struct{})},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 1,
+		},
+		{
+			Name:                   "unsupported-array-values",
+			MaxContainerLength:     2,
+			Data:                   []interface{}{"supported", func() {}, "supported", make(chan struct{})},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 2,
+		},
+		{
+			Name:                   "unsupported-array-values",
+			MaxContainerLength:     1,
+			Data:                   []interface{}{func() {}, "supported", make(chan struct{}), "supported"},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 1,
+		},
+		{
+			Name:                   "unsupported-array-values",
+			MaxContainerLength:     2,
+			Data:                   []interface{}{func() {}, "supported", make(chan struct{}), "supported"},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 2,
+		},
+		{
+			Name:                   "unsupported-array-values",
+			MaxContainerLength:     1,
+			Data:                   []interface{}{func() {}, make(chan struct{}), "supported"},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 1,
+		},
+		{
+			Name:                   "unsupported-array-values",
+			MaxContainerLength:     3,
+			Data:                   []interface{}{"supported", func() {}, make(chan struct{})},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 1,
+		},
+		{
+			Name:                   "unsupported-array-values",
+			MaxContainerLength:     3,
+			Data:                   []interface{}{func() {}, "supported", make(chan struct{})},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 1,
+		},
+		{
+			Name:                   "unsupported-array-values",
+			MaxContainerLength:     3,
+			Data:                   []interface{}{func() {}, make(chan struct{}), "supported"},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 1,
+		},
+		{
+			Name:                   "unsupported-array-values",
+			MaxContainerLength:     2,
+			Data:                   []interface{}{func() {}, make(chan struct{}), "supported", "supported"},
+			ExpectedWAFValueType:   wafArrayType,
+			ExpectedWAFValueLength: 2,
+		},
+		{
+			Name: "unsupported-map-key-types",
+			Data: map[interface{}]int{
+				"supported":           1,
+				interface{ m() }(nil): 1,
+				nil:                   1,
+				(*int)(nil):           1,
+				(*string)(nil):        1,
+				make(chan struct{}):   1,
+			},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 1,
+		},
+		{
+			Name: "unsupported-map-key-types",
+			Data: map[interface{}]int{
+				interface{ m() }(nil): 1,
+				nil:                   1,
+				(*int)(nil):           1,
+				(*string)(nil):        1,
+				make(chan struct{}):   1,
+			},
+			ExpectedWAFValueType:   wafMapType,
+			ExpectedWAFValueLength: 0,
+		},
+		{
+			Name: "unsupported-map-values",
+			Data: map[string]interface{}{
+				"k0": "supported",
+				"k1": func() {},
+				"k2": make(chan struct{}),
+			},
+			MaxContainerLength:     3,
+			ExpectedWAFValueLength: 1,
+			ExpectedWAFValueType:   wafMapType,
+		},
+		{
+			Name: "unsupported-map-values",
+			Data: map[string]interface{}{
+				"k0": "supported",
+				"k1": "supported",
+				"k2": make(chan struct{}),
+			},
+			MaxContainerLength:     3,
+			ExpectedWAFValueLength: 2,
+			ExpectedWAFValueType:   wafMapType,
+		},
+		{
+			Name: "unsupported-map-values",
+			Data: map[string]interface{}{
+				"k0": "supported",
+				"k1": "supported",
+				"k2": make(chan struct{}),
+			},
+			MaxContainerLength:     1,
+			ExpectedWAFValueLength: 1,
+			ExpectedWAFValueType:   wafMapType,
+		},
+		{
+			Name: "unsupported-struct-values",
+			Data: struct {
+				F0 string
+				F1 func()
+				F2 chan struct{}
+			}{
+				F0: "supported",
+				F1: func() {},
+				F2: make(chan struct{}),
+			},
+			MaxContainerLength:     3,
+			ExpectedWAFValueLength: 1,
+			ExpectedWAFValueType:   wafMapType,
+		},
+		{
+			Name: "unsupported-map-values",
+			Data: struct {
+				F0 string
+				F1 string
+				F2 chan struct{}
+			}{
+				F0: "supported",
+				F1: "supported",
+				F2: make(chan struct{}),
+			},
+			MaxContainerLength:     3,
+			ExpectedWAFValueLength: 2,
+			ExpectedWAFValueType:   wafMapType,
+		},
+		{
+			Name: "unsupported-map-values",
+			Data: struct {
+				F0 string
+				F1 string
+				F2 chan struct{}
+			}{
+				F0: "supported",
+				F1: "supported",
+				F2: make(chan struct{}),
+			},
+			MaxContainerLength:     1,
+			ExpectedWAFValueLength: 1,
+			ExpectedWAFValueType:   wafMapType,
+		},
+	} {
+		tc := tc
+		t.Run(tc.Name, func(t *testing.T) {
+			maxValueDepth := 10
+			if max := tc.MaxValueDepth; max != nil {
+				maxValueDepth = max.(int)
+			}
+			maxContainerLength := 1000
+			if max := tc.MaxContainerLength; max != nil {
+				maxContainerLength = max.(int)
+			}
+			maxStringLength := 4096
+			if max := tc.MaxStringLength; max != nil {
+				maxStringLength = max.(int)
+			}
+			e := encoder{
+				objectMaxDepth:   maxValueDepth,
+				stringMaxSize:    maxStringLength,
+				containerMaxSize: maxContainerLength,
+			}
+			wo, err := e.Encode(tc.Data)
+			if tc.ExpectedError != nil {
+				require.Error(t, err)
+				require.Equal(t, tc.ExpectedError, err)
+				require.Nil(t, wo)
+				return
+			}
 
-	encoder := newMaxEncoder()
-	wo, err := encoder.Encode(nilMap)
-	assert.NoError(t, err)
+			require.NoError(t, err)
+			require.NotEqual(t, &wafObject{}, wo)
 
-	assert.True(t, wo.isMap())
-	assert.Equal(t, uint64(0), wo.nbEntries)
+			if tc.ExpectedWAFValueType != 0 {
+				require.Equal(t, tc.ExpectedWAFValueType, int(wo._type), "bad waf value type")
+			}
+			if tc.ExpectedWAFValueLength != 0 {
+				require.Equal(t, tc.ExpectedWAFValueLength, int(wo.nbEntries), "bad waf value length")
+			}
+			require.Equal(t, tc.ExpectedWAFValueType, int(wo._type), "bad waf string value type")
+
+			switch tc.ExpectedWAFValueType {
+			case wafIntType:
+				require.Equal(t, tc.ExpectedWafValueInt, int64(wo.value))
+			case wafUintType:
+				require.Equal(t, tc.ExpectedWafValueUint, uint64(wo.value))
+			case wafFloatType:
+				require.Equal(t, tc.ExpectedWafValueFloat, uintptrToNative[float64](wo.value))
+			case wafBoolType:
+				require.Equal(t, tc.ExpectedWafValueBool, uintptrToNative[bool](wo.value))
+			}
+
+			if expectedStr := tc.ExpectedWAFString; expectedStr != "" {
+				cbuf := wo.value
+				gobuf := []byte(expectedStr)
+				require.Equal(t, len(gobuf), int(wo.nbEntries), "bad waf value length")
+				for i, gobyte := range gobuf {
+					// Go pointer arithmetic for cbyte := cbuf[i]
+					cbyte := *(*uint8)(unsafe.Pointer(cbuf + uintptr(i))) //nolint:govet
+					if cbyte != gobyte {
+						t.Fatalf("bad waf string value content: i=%d cbyte=%d gobyte=%d", i, cbyte, gobyte)
+					}
+				}
+			}
+
+			// Pass the encoded value to the WAF to make sure it doesn't return an error
+			waf, err := newDefaultHandle(newArachniTestRule([]ruleInput{{Address: "my.input"}}, nil))
+			require.NoError(t, err)
+			defer waf.Close()
+			wafCtx := NewContext(waf)
+			require.NotNil(t, wafCtx)
+			defer wafCtx.Close()
+			_, err = wafCtx.Run(RunAddressData{Persistent: map[string]interface{}{
+				"my.input": tc.Data,
+			}}, time.Second)
+			require.NoError(t, err)
+		})
+	}
 }
